# api/openapi.json
{"openapi":"3.1.0","info":{"title":"FeedbackIQ API","description":"Customer feedback intelligence platform API. Classify reviews, generate AI insights, download reports.","version":"1.0.0"},"paths":{"/auth/signup":{"post":{"tags":["Authentication"],"summary":"Register a new user account","operationId":"signup_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Authentication"],"summary":"Login with email and password","operationId":"login_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/refresh":{"post":{"tags":["Authentication"],"summary":"Refresh access token using refresh token","operationId":"refresh_token_auth_refresh_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["Authentication"],"summary":"Get current user profile and stats","operationId":"get_me_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserMeResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/auth/profile":{"get":{"tags":["Authentication"],"summary":"Get saved company profile","operationId":"get_profile_auth_profile_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"put":{"tags":["Authentication"],"summary":"Save or update company profile","operationId":"update_profile_auth_profile_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileUpdateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/auth/change-password":{"post":{"tags":["Authentication"],"summary":"Change account password","operationId":"change_password_auth_change_password_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/auth/history":{"get":{"tags":["Authentication"],"summary":"Get analysis session history","operationId":"get_history_auth_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/sessions":{"get":{"tags":["Sessions"],"summary":"List current user's analysis sessions","operationId":"list_sessions_sessions_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionsListResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["Sessions"],"summary":"Create a new analysis session","operationId":"create_session_sessions_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSessionRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSessionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/analyse/text":{"post":{"tags":["Analysis"],"summary":"Analyse pasted review text","operationId":"analyse_text_analyse_text_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_analyse_text_analyse_text_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalyseResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/analyse/file":{"post":{"tags":["Analysis"],"summary":"Analyse uploaded CSV or Excel file","operationId":"analyse_file_analyse_file_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_analyse_file_analyse_file_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalyseResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/dashboard/{session_id}":{"get":{"tags":["Dashboard"],"summary":"Get full dashboard data for a session","operationId":"get_dashboard_dashboard__session_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DashboardResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/action-plan/{session_id}":{"post":{"tags":["Action Plan"],"summary":"Generate AI action plan for a completed session","operationId":"create_action_plan_action_plan__session_id__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActionPlanResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/report/{session_id}":{"get":{"tags":["Report"],"summary":"Download PDF report for a session","operationId":"download_report_report__session_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/export/{session_id}":{"get":{"tags":["Export"],"summary":"Download classified results as CSV","operationId":"export_csv_export__session_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"string","title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["System"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}}},"components":{"schemas":{"ActionPlanResponse":{"properties":{"session_id":{"type":"string","title":"Session Id"},"success":{"type":"boolean","title":"Success"},"result":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Result"},"health_score":{"type":"integer","title":"Health Score"},"health_label":{"type":"string","title":"Health Label"},"provider":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Provider"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["session_id","success","result","health_score","health_label","provider","error"],"title":"ActionPlanResponse"},"AnalyseResponse":{"properties":{"session_id":{"type":"string","title":"Session Id"},"total_classified":{"type":"integer","title":"Total Classified"},"total_failed":{"type":"integer","title":"Total Failed"},"gemini_fallback_count":{"type":"integer","title":"Gemini Fallback Count"},"failed_batches":{"items":{},"type":"array","title":"Failed Batches"},"preprocessing":{"$ref":"#/components/schemas/PreprocessingSummary"}},"type":"object","required":["session_id","total_classified","total_failed","gemini_fallback_count","failed_batches","preprocessing"],"title":"AnalyseResponse"},"Body_analyse_file_analyse_file_post":{"properties":{"session_id":{"type":"string","title":"Session Id"},"file":{"type":"string","format":"binary","title":"File"},"column":{"type":"string","title":"Column","default":""}},"type":"object","required":["session_id","file"],"title":"Body_analyse_file_analyse_file_post"},"Body_analyse_text_analyse_text_post":{"properties":{"session_id":{"type":"string","title":"Session Id"},"raw_text":{"type":"string","title":"Raw Text"}},"type":"object","required":["session_id","raw_text"],"title":"Body_analyse_text_analyse_text_post"},"Body_login_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_auth_login_post"},"ChangePasswordRequest":{"properties":{"current_password":{"type":"string","title":"Current Password"},"new_password":{"type":"string","maxLength":128,"minLength":8,"title":"New Password"}},"type":"object","required":["current_password","new_password"],"title":"ChangePasswordRequest"},"CreateSessionRequest":{"properties":{"company_name":{"type":"string","maxLength":100,"minLength":1,"title":"Company Name"},"industry":{"type":"string","minLength":1,"title":"Industry"},"categories":{"items":{"type":"string"},"type":"array","maxItems":8,"minItems":2,"title":"Categories"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","default":""},"urgency_definition":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Urgency Definition","default":""}},"type":"object","required":["company_name","industry","categories"],"title":"CreateSessionRequest"},"CreateSessionResponse":{"properties":{"session_id":{"type":"string","title":"Session Id"},"profile":{"additionalProperties":true,"type":"object","title":"Profile"},"created_at":{"type":"string","title":"Created At"},"user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id"}},"type":"object","required":["session_id","profile","created_at"],"title":"CreateSessionResponse"},"DashboardResponse":{"properties":{"session_id":{"type":"string","title":"Session Id"},"profile":{"additionalProperties":true,"type":"object","title":"Profile"},"dashboard_data":{"additionalProperties":true,"type":"object","title":"Dashboard Data"},"classification_done":{"type":"boolean","title":"Classification Done"},"total_classified":{"type":"integer","title":"Total Classified"}},"type":"object","required":["session_id","profile","dashboard_data","classification_done","total_classified"],"title":"DashboardResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"version":{"type":"string","title":"Version"},"timestamp":{"type":"string","title":"Timestamp"},"modules":{"items":{"type":"string"},"type":"array","title":"Modules"}},"type":"object","required":["status","version","timestamp","modules"],"title":"HealthResponse"},"PreprocessingSummary":{"properties":{"input_count":{"type":"integer","title":"Input Count"},"final_count":{"type":"integer","title":"Final Count"},"noise_removed":{"type":"integer","title":"Noise Removed"},"exact_duplicates_removed":{"type":"integer","title":"Exact Duplicates Removed"},"near_duplicates_removed":{"type":"integer","title":"Near Duplicates Removed"},"short_removed":{"type":"integer","title":"Short Removed"}},"type":"object","required":["input_count","final_count","noise_removed","exact_duplicates_removed","near_duplicates_removed","short_removed"],"title":"PreprocessingSummary"},"ProfileUpdateRequest":{"properties":{"company_name":{"type":"string","maxLength":100,"minLength":1,"title":"Company Name"},"industry":{"type":"string","minLength":1,"title":"Industry"},"categories":{"items":{"type":"string"},"type":"array","maxItems":8,"minItems":2,"title":"Categories"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","default":""},"urgency_definition":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Urgency Definition","default":""}},"type":"object","required":["company_name","industry","categories"],"title":"ProfileUpdateRequest"},"RefreshRequest":{"properties":{"refresh_token":{"type":"string","title":"Refresh Token"}},"type":"object","required":["refresh_token"],"title":"RefreshRequest"},"SessionSummary":{"properties":{"session_id":{"type":"string","title":"Session Id"},"label":{"type":"string","title":"Label"},"created_at":{"type":"string","title":"Created At"},"total_reviews":{"type":"integer","title":"Total Reviews"},"overall_score":{"type":"number","title":"Overall Score"}},"type":"object","required":["session_id","label","created_at","total_reviews","overall_score"],"title":"SessionSummary"},"SessionsListResponse":{"properties":{"sessions":{"items":{"$ref":"#/components/schemas/SessionSummary"},"type":"array","title":"Sessions"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["sessions","total"],"title":"SessionsListResponse"},"SignupRequest":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","maxLength":128,"minLength":8,"title":"Password"},"full_name":{"type":"string","maxLength":100,"minLength":1,"title":"Full Name"}},"type":"object","required":["email","password","full_name"],"title":"SignupRequest"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"user_id":{"type":"string","title":"User Id"},"email":{"type":"string","title":"Email"},"full_name":{"type":"string","title":"Full Name"},"has_profile":{"type":"boolean","title":"Has Profile"}},"type":"object","required":["access_token","refresh_token","user_id","email","full_name","has_profile"],"title":"TokenResponse"},"UserMeResponse":{"properties":{"user_id":{"type":"string","title":"User Id"},"email":{"type":"string","title":"Email"},"full_name":{"type":"string","title":"Full Name"},"created_at":{"type":"string","title":"Created At"},"last_login":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Login"},"profile":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Profile"},"session_count":{"type":"integer","title":"Session Count"},"has_profile":{"type":"boolean","title":"Has Profile"}},"type":"object","required":["user_id","email","full_name","created_at","last_login","profile","session_count","has_profile"],"title":"UserMeResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/auth/login"}}}}}}